simcompare refactored into SV and SC utilities
